navigator: fold CompassPoint into to-compass-point, move out of base/
CompassPoint was a UI-shaped string union sitting in
controllers/types.ts despite having no controller relationship.
to-compass-point.ts was the only file in base/. Inline the type into
to-compass-point.ts and move the file to util/. Drop the base/
directory.
